+++ src/app/components/filesystem-browser/tag-recalled.js
@@ -10,6 +10,8 @@
 
 import Component from '@ember/component';
 import I18n from 'onedata-gui-common/mixins/components/i18n';
+import { reads } from '@ember/object/computed';
+import notImplementedIgnore from 'onedata-gui-common/utils/not-implemented-ignore';
 
 export default Component.extend(I18n, {
   tagName: '',
@@ -30,4 +32,21 @@ export default Component.extend(I18n, {
    * @type {Boolean}
    */
   disabled: false,
+
+  /**
+   * @virtual optional
+   * @type () => void
+   */
+  onTagClicked: notImplementedIgnore,
+
+  errorOccurred: reads('file.archiveRecallState.errorOccurred'),
+
+  actions: {
+    tagClicked() {
+      const onTagClicked = this.get('onTagClicked');
+      if (onTagClicked) {
+        onTagClicked();
+      }
+    },
+  },
 });