+++ src/app/locales/en/components/archive-recall.js
@@ -1,8 +1,13 @@
 export default {
   archiveRecallStartSuccess: 'Archive recall process successfully started.',
   archiveRecallProcessStart: 'archive recall process start',
+  fileType: {
+    file: 'file',
+    dir: 'directory',
+    symlink: 'symbolic link',
+  },
   browserValidation: {
-    recalling: 'Cannot recall here – selected is currently being recalled',
+    recalling: 'Cannot recall here – selected directory is currently being recalled',
   },
   targetNameValidation: {
     exists: 'This filename is already in use',