+++ src/app/locales/en/translations.js
@@ -33,6 +33,7 @@ import itemsSelectBrowser from './components/items-select-browser';
 import datasetProtection from './components/dataset-protection';
 import filePath from './components/file-path';
 import archiveRecall from './components/archive-recall';
+import recallInfoModal from './components/recall-info-modal';
 
 import uploadManager from './services/upload-manager';
 
@@ -80,6 +81,7 @@ const translations = {
     datasetProtection,
     filePath,
     archiveRecall,
+    recallInfoModal,
   },
   services: {
     uploadManager,