+++ src/app/models/archive-recall-state.js
@@ -9,6 +9,10 @@ import Model from 'ember-data/model';
 import attr from 'ember-data/attr';
 import StaticGraphModelMixin from 'onedata-gui-websocket-client/mixins/models/static-graph-model';
 import GraphSingleModelMixin from 'onedata-gui-websocket-client/mixins/models/graph-single-model';
+import { or } from 'ember-awesome-macros';
+import { get } from '@ember/object';
+import computedPipe from 'onedata-gui-common/utils/ember/computed-pipe';
+import _ from 'lodash';
 
 // model name differs from aspect name to avoid "s" on end of model name
 export const aspect = 'archive_recall_progress';
@@ -22,4 +26,27 @@ export default Model.extend(GraphSingleModelMixin, {
    * @type {ComputedProperty<{ fileId: String, reason: Object }>}
    */
   lastError: attr('object', { defaultValue: null }),
+
+  errorOccurred: or(
+    'filesFailed',
+    computedPipe('lastError', (value) => !_.isEmpty(value))
+  ),
+
+  /**
+   *
+   * @param {Models.ArchiveRecalInfo} archiveRecallInfo info associated with the same
+   *   recall process
+   * @returns {Boolean}
+   */
+  isFinished(archiveRecallInfo) {
+    if (!archiveRecallInfo) {
+      return;
+    }
+    const {
+      filesCopied,
+      filesFailed,
+    } = this.getProperties('filesCopied', 'filesFailed');
+    const totalFileCount = get(archiveRecallInfo, 'totalFileCount');
+    return filesCopied + filesFailed >= totalFileCount;
+  },
 }).reopenClass(StaticGraphModelMixin);