+++ src/app/utils/archive-recall-state-watcher.js
@@ -10,8 +10,13 @@
 import EmberObject, { get } from '@ember/object';
 import { bool } from 'ember-awesome-macros';
 import Looper from 'onedata-gui-common/utils/looper';
+import { inject as service } from '@ember/service';
+import OwnerInjector from 'onedata-gui-common/mixins/owner-injector';
+import { all as allSettled } from 'rsvp';
 
-export default EmberObject.extend({
+export default EmberObject.extend(OwnerInjector, {
+  fileManager: service(),
+
   /**
    * @virtual
    * @type {File}
@@ -26,10 +31,17 @@ export default EmberObject.extend({
   interval: 2000,
 
   /**
-   * Created on init.
+   * Created on `start`.
    * @type {Looper}
    */
   looper: null,
+
+  /**
+   * Stores entity IDs of directories to be auto-refreshed on update.
+   * Created on init.
+   * @type {Set<String>}
+   */
+  refreshedDirsIdSet: null,
 
   /**
    * @type {Boolean}
@@ -45,6 +57,11 @@ export default EmberObject.extend({
    * @type {'state'|'info'}
    */
   pollingMode: 'state',
+
+  init() {
+    this._super(...arguments);
+    this.set('refreshedDirsIdSet', new Set());
+  },
 
   /**
    * @override
@@ -93,7 +110,7 @@ export default EmberObject.extend({
       this.stop();
       return;
     }
-    isFinished = get(info, 'finishTime');
+    isFinished = Boolean(get(info, 'finishTime'));
     if (isFinished) {
       this.stop();
       return;
@@ -111,7 +128,7 @@ export default EmberObject.extend({
     }
     // pollingMode could change if reloadState failed, so check one more time
     if (pollingMode === 'state') {
-      isFinished = isFinished || get(state, 'bytesCopied') >= get(info, 'totalByteSize');
+      isFinished = isFinished || state.isFinished(info);
       shouldUpdateInfo = (
         !get(info, 'startTime') && get(state, 'bytesCopied')
       ) || isFinished;
@@ -130,8 +147,36 @@ export default EmberObject.extend({
         return;
       }
     }
+    this.updateFileBrowsers();
     if (isFinished) {
       this.stop();
+    }
+  },
+
+  addToAutoRefresh(file) {
+    const refreshedDirsIdSet = this.get('refreshedDirsIdSet');
+    const dirToRefreshId = this.getDirIdToRefresh(file);
+    refreshedDirsIdSet.add(dirToRefreshId);
+  },
+
+  /**
+   * @param {Models.File} file
+   * @returns {String}
+   */
+  getDirIdToRefresh(file) {
+    return get(file, 'type') === 'dir' ?
+      get(file, 'entityId') : file.relationEntityId('parent');
+  },
+
+  async updateFileBrowsers() {
+    const {
+      refreshedDirsIdSet,
+      fileManager,
+    } = this.getProperties('refreshedDirsIdSet', 'fileManager');
+    if (refreshedDirsIdSet.size) {
+      await allSettled([...refreshedDirsIdSet].map(dirId =>
+        fileManager.dirChildrenRefresh(dirId)
+      ));
     }
   },