+++ src/tests/integration/components/filesystem-browser/file-features-test.js
@@ -1,33 +1,38 @@
 import { expect } from 'chai';
-import { describe, it } from 'mocha';
+import { describe, it, beforeEach } from 'mocha';
 import { setupComponentTest } from 'ember-mocha';
 import hbs from 'htmlbars-inline-precompile';
 import sinon from 'sinon';
 import { click } from 'ember-native-dom-helpers';
 import { createArchiveRecallData } from '../../../helpers/archive-recall';
 import wait from 'ember-test-helpers/wait';
+import { lookupService } from '../../../helpers/stub-service';
 
 describe('Integration | Component | filesystem browser/file features', function () {
   setupComponentTest('filesystem-browser/file-features', {
     integration: true,
   });
 
-  ['none', 'direct'].forEach(membership => {
-    it(`does not show collapsed inherited tag if features are "${membership}" in collapsed mode`, function () {
-      const item = {
-        effDatasetMembership: membership,
-        effQosMembership: membership,
-        recallingMembership: membership,
-      };
-      this.set('item', item);
+  beforeEach(function () {
+    this.createItem = (...args) => createFileItem(this, ...args);
+  });
 
-      this.render(hbs `{{filesystem-browser/file-features
+  ['none', 'direct'].forEach(membership => {
+    it(`does not show collapsed inherited tag if features are "${membership}" in collapsed mode`,
+      async function () {
+        await this.createItem({
+          effDatasetMembership: membership,
+          effQosMembership: membership,
+          recallingMembership: membership,
+        });
+
+        this.render(hbs `{{filesystem-browser/file-features
         item=item
         initiallyExpanded=false
       }}`);
 
-      expect(this.$('.file-status-inherited-collapsed')).to.not.exist;
-    });
+        expect(this.$('.file-status-inherited-collapsed')).to.not.exist;
+      });
   });
 
   ['ancestor', 'directAndAncestor'].forEach(membership => {
@@ -41,11 +46,10 @@ describe('Integration | Component | filesystem browser/file features', function 
         return;
       }
       it(`shows collapsed inherited tag if "${feature}" feature is "${membership}" in collapsed mode`,
-        function () {
-          const item = {
+        async function () {
+          await this.createItem({
             [feature]: membership,
-          };
-          this.set('item', item);
+          });
 
           this.render(hbs `{{filesystem-browser/file-features
             item=item
@@ -57,13 +61,12 @@ describe('Integration | Component | filesystem browser/file features', function 
     });
   });
 
-  it('shows tags with "direct" features in expanded mode', function () {
-    const item = {
+  it('shows tags with "direct" features in expanded mode', async function () {
+    await this.createItem({
       effDatasetMembership: 'direct',
       effQosMembership: 'direct',
       recallingMembership: 'direct',
-    };
-    this.set('item', item);
+    });
 
     this.render(hbs `{{filesystem-browser/file-features
       item=item
@@ -76,13 +79,12 @@ describe('Integration | Component | filesystem browser/file features', function 
   });
 
   ['direct', 'directAndAncestor'].forEach(membership => {
-    it(`shows tags with "${membership}" features in collapsed mode`, function () {
-      const item = {
+    it(`shows tags with "${membership}" features in collapsed mode`, async function () {
+      await this.createItem({
         effDatasetMembership: membership,
         effQosMembership: membership,
         recallingMembership: membership,
-      };
-      this.set('item', item);
+      });
 
       this.render(hbs `{{filesystem-browser/file-features
         item=item
@@ -97,12 +99,11 @@ describe('Integration | Component | filesystem browser/file features', function 
 
   // NOTE: not testing recalling tag, because it's not using "directAndAncestor"
   it('shows direct tags and collapsed inheritance icon when features are "directAndAncestor" in collapsed mode',
-    function () {
-      const item = {
+    async function () {
+      await this.createItem({
         effDatasetMembership: 'directAndAncestor',
         effQosMembership: 'directAndAncestor',
-      };
-      this.set('item', item);
+      });
 
       this.render(hbs `{{filesystem-browser/file-features
         item=item
@@ -121,12 +122,11 @@ describe('Integration | Component | filesystem browser/file features', function 
 
   // NOTE: not testing recalling tag, because it's not using "directAndAncestor"
   it('shows pill-like direct-ancestor tags without collapsed inheritance icon when features are "directAndAncestor" in expanded mode',
-    function () {
-      const item = {
+    async function () {
+      await this.createItem({
         effDatasetMembership: 'directAndAncestor',
         effQosMembership: 'directAndAncestor',
-      };
-      this.set('item', item);
+      });
 
       this.render(hbs `{{filesystem-browser/file-features
         item=item
@@ -148,13 +148,12 @@ describe('Integration | Component | filesystem browser/file features', function 
   );
 
   it('shows feature ancestor tags without collapsed inheritance icon when features are "ancestor" in expanded mode',
-    function () {
-      const item = {
+    async function () {
+      await this.createItem({
         effDatasetMembership: 'ancestor',
         effQosMembership: 'ancestor',
         recallingMembership: 'ancestor',
-      };
-      this.set('item', item);
+      });
 
       this.render(hbs `{{filesystem-browser/file-features
         item=item
@@ -182,17 +181,16 @@ describe('Integration | Component | filesystem browser/file features', function 
   ].forEach(({ tag, action }) => {
     it(`invokes onInvokeItemAction item and actionName="${action}" when clicking on "${tag}" tag`,
       async function () {
-        const item = {
+        const item = await this.createItem({
           effDatasetMembership: 'direct',
           effQosMembership: 'direct',
-        };
+        });
         const onInvokeItemAction = sinon.spy();
         const spacePrivileges = {
           view: true,
           viewQos: true,
         };
         this.setProperties({
-          item,
           onInvokeItemAction,
           spacePrivileges,
         });
@@ -218,15 +216,14 @@ describe('Integration | Component | filesystem browser/file features', function 
     { tag: 'qos', text: 'QoS' },
     { tag: 'recalling', text: 'Recalling' },
   ].forEach(({ tag, text }) => {
-    it(`displays "${text}" text on ${tag} tag`, function () {
-      const item = {
+    it(`displays "${text}" text on ${tag} tag`, async function () {
+      await this.createItem({
         effDatasetMembership: 'direct',
         effQosMembership: 'direct',
         recallingMembership: 'direct',
-      };
+      });
       const onInvokeItemAction = sinon.spy();
       this.setProperties({
-        item,
         onInvokeItemAction,
       });
 
@@ -241,6 +238,31 @@ describe('Integration | Component | filesystem browser/file features', function 
       expect($tag.text().trim()).to.contain(text);
     });
   });
+
+  // NOTE: "directAndAncestor" not used with recalling feature
+  it('changes direct tags into direct-ancestor tags after inheritance tag click', async function () {
+    await this.createItem({
+      effDatasetMembership: 'directAndAncestor',
+      effQosMembership: 'directAndAncestor',
+    });
+
+    this.render(hbs `{{filesystem-browser/file-features
+      item=item
+      initiallyExpanded=false
+    }}`);
+    const $inheritanceTag = this.$('.file-status-inherited-collapsed');
+    expect($inheritanceTag).to.have.length(1);
+    await click($inheritanceTag[0]);
+
+    expect(this.$('.file-status-inherited-collapsed'), 'inheritance tag after click')
+      .to.not.exist;
+    expect(this.$('.dataset-file-status-tag-group .inherited-icon'), 'dataset inherited a. click')
+      .to.exist;
+    expect(this.$('.qos-file-status-tag-group .inherited-icon'), 'qos inherited a. click')
+      .to.exist;
+  });
+
+  //#region recalling tag
 
   it('displays percentage progress of archive recalling in recalling tag', async function () {
     createArchiveRecallData(this);
@@ -268,27 +290,36 @@ describe('Integration | Component | filesystem browser/file features', function 
     expect(text).to.contain('50%');
   });
 
-  // NOTE: "directAndAncestor" not used with recalling feature
-  it('changes direct tags into direct-ancestor tags after inheritance tag click', async function () {
-    const item = {
-      effDatasetMembership: 'directAndAncestor',
-      effQosMembership: 'directAndAncestor',
-    };
-    this.set('item', item);
+  it('has percentage width style applied to progress element in recalling tag', async function () {
+    createArchiveRecallData(this);
+    const targetFile = this.get('targetFile');
+    this.set(
+      'archiveRecallState.bytesCopied',
+      this.get('archiveRecallInfo.totalByteSize') / 2
+    );
+    const onInvokeItemAction = sinon.spy();
+    this.setProperties({
+      item: targetFile,
+      onInvokeItemAction,
+    });
 
     this.render(hbs `{{filesystem-browser/file-features
       item=item
       initiallyExpanded=false
+      onInvokeItemAction=onInvokeItemAction
     }}`);
-    const $inheritanceTag = this.$('.file-status-inherited-collapsed');
-    expect($inheritanceTag).to.have.length(1);
-    await click($inheritanceTag[0]);
+    await wait();
 
-    expect(this.$('.file-status-inherited-collapsed'), 'inheritance tag after click')
-      .to.not.exist;
-    expect(this.$('.dataset-file-status-tag-group .inherited-icon'), 'dataset inherited a. click')
-      .to.exist;
-    expect(this.$('.qos-file-status-tag-group .inherited-icon'), 'qos inherited a. click')
-      .to.exist;
+    const tagProgress = this.$('.file-status-recalling .tag-progress')[0];
+    expect(tagProgress.style.width).to.equal('50%');
   });
+
+  //#endregion
 });
+
+async function createFileItem(testCase, data) {
+  const store = lookupService(testCase, 'store');
+  const file = store.createRecord('file', data);
+  await file.save();
+  return testCase.set('item', file);
+}